Allow arbitrary xpub account imports.
This commit implements a new JSON-RPC method "importxpub" which
records an arbitrary HD extended pubkey to the wallet database and
allows new addresses to be derived for the account.  It takes two
positional parameters: an account name and the HD extended pubkey in
string format.

This change contains a database upgrade and it is not possible to
downgrade a wallet database once run.  Reverting to older code
requires restoring the database from backup or seed restoring.